python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
marker-pdf==0.2.17
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
